move postman package position
